ok, listo para p3-2

+++ src/main/java/distance/EuclideanDistance.java
@@ -7,12 +7,12 @@ import java.util.List;
 public class EuclideanDistance implements Distance {
     @Override
     public double calculateDistance(List<Double> p, List<Double> q) {
-        Double sumatorio = 0.0;
+        double sumatorio = 0.0;
         for (int i = 0; i < p.size(); i++) {
-            Double resta = (p.get(i) - q.get(i));
+            double resta = (p.get(i) - q.get(i));
             resta = resta * resta;
             sumatorio += resta;
         }
-        return (Double) Math.sqrt(sumatorio);
+        return Math.sqrt(sumatorio);
     }
 }
